[SofaKernel] Add new DataTypeInfo and  remove the fallback syste.
