+++ SofaKernel/modules/SofaBaseMechanics/BarycentricMappers/BarycentricMapper.h
@@ -27,6 +27,8 @@
 #include <SofaBaseLinearSolver/CompressedRowSparseMatrix.h>
 #include <sofa/core/topology/BaseMeshTopology.h>
 #include <sofa/defaulttype/typeinfo/DataTypeInfo_Vec.h>
+#include <sofa/defaulttype/typeinfo/DataTypeInfo_RigidTypes.h>
+
 namespace sofa
 {
