+++ SofaKernel/modules/SofaBaseMechanics/BarycentricMappers/BarycentricMapper.inl
@@ -23,6 +23,7 @@
 #define SOFA_COMPONENT_MAPPING_BARYCENTRICMAPPER_INL
 
 #include "BarycentricMapper.h"
+#include <sofa/defaulttype/typeinfo/DataTypeInfo_Scalar.h>
 
 namespace sofa
 {
