+++ SofaKernel/modules/SofaBaseMechanics/BarycentricMappers/BarycentricMapperEdgeSetTopology.cpp
@@ -22,6 +22,8 @@
 #ifndef SOFA_COMPONENT_MAPPING_BARYCENTRICMAPPEREDGESETTOPOLOGY_CPP
 #define SOFA_COMPONENT_MAPPING_BARYCENTRICMAPPEREDGESETTOPOLOGY_CPP
 
+#include <sofa/defaulttype/typeinfo/DataTypeInfo_Scalar.h>
+#include <sofa/defaulttype/typeinfo/DataTypeInfo_Vec.h>
 #include "BarycentricMapperEdgeSetTopology.inl"
 
 namespace sofa
