+++ SofaKernel/modules/SofaBaseMechanics/BarycentricMappers/BarycentricMapperTriangleSetTopology.cpp
@@ -21,6 +21,10 @@
 ******************************************************************************/
 #ifndef SOFA_COMPONENT_MAPPING_BARYCENTRICMAPPERTRIANGLESETTOPOLOGY_CPP
 #define SOFA_COMPONENT_MAPPING_BARYCENTRICMAPPERTRIANGLESETTOPOLOGY_CPP
+
+#include <sofa/defaulttype/typeinfo/DataTypeInfo_Scalar.h>
+#include <sofa/defaulttype/typeinfo/DataTypeInfo_Vec.h>
+
 #include "BarycentricMapperTriangleSetTopology.inl"
 
 namespace sofa
