+++ SofaKernel/modules/SofaBaseMechanics/BarycentricMappers/TopologyBarycentricMapper.inl
@@ -22,6 +22,7 @@
 #ifndef SOFA_COMPONENT_MAPPING_TOPOLOGYBARYCENTRICMAPPER_INL
 #define SOFA_COMPONENT_MAPPING_TOPOLOGYBARYCENTRICMAPPER_INL
 
+#include <sofa/defaulttype/typeinfo/DataTypeInfo_Scalar.h>
 #include "TopologyBarycentricMapper.h"
 
 namespace sofa
