+++ SofaKernel/modules/SofaBaseMechanics/IdentityMapping.h
@@ -30,6 +30,7 @@
 #include <vector>
 #include <memory>
 #include <SofaEigen2Solver/EigenSparseMatrix.h>
+#include <sofa/defaulttype/typeinfo/DataTypeInfo_RigidTypes.h>
 
 
 namespace sofa
