+++ SofaKernel/modules/SofaBaseMechanics/IdentityMapping.inl
@@ -24,6 +24,8 @@
 
 #include <SofaBaseMechanics/IdentityMapping.h>
 #include <sofa/defaulttype/typeinfo/DataTypeInfo_Vec.h>
+#include <sofa/defaulttype/typeinfo/DataTypeInfo_Scalar.h>
+
 #include <sofa/defaulttype/RigidTypes.h>
 #include <sofa/defaulttype/VecTypes.h>
