+++ SofaKernel/modules/SofaDefaultType/src/sofa/defaulttype/Quat.h
@@ -23,33 +23,17 @@
 #define SOFA_DEFAULTTYPE_QUAT_H
 
 #include <sofa/helper/Quater.h>
-#include <sofa/defaulttype/DataTypeInfo.h>
 
 namespace sofa
 {
 
 namespace defaulttype
 {
+
 typedef helper::Quater<double> Quatd; ///< alias
 typedef helper::Quater<float>  Quatf; ///< alias
 typedef helper::Quater<SReal>  Quat; ///< alias
 typedef Quat Quaternion; ///< alias
-
-// Specialization of the defaulttype::DataTypeInfo type traits template
-
-template<class T>
-struct DataTypeInfo< sofa::helper::Quater<T> > : public FixedArrayTypeInfo< sofa::helper::Quater<T> >
-{
-    static std::string name() { std::ostringstream o; o << "Quater<" << DataTypeName<T>::name() << ">"; return o.str(); }
-};
-
-// The next line hides all those methods from the doxygen documentation
-/// \cond TEMPLATE_OVERRIDES
-
-template<> struct DataTypeName<defaulttype::Quatf> { static const char* name() { return "Quatf"; } };
-template<> struct DataTypeName<defaulttype::Quatd> { static const char* name() { return "Quatd"; } };
-
-/// \endcond
 
 } // namespace defaulttype
