+++ SofaKernel/modules/SofaRigid/RigidMapping.h
@@ -31,6 +31,7 @@
 
 #include <sofa/defaulttype/VecTypes.h>
 #include <sofa/defaulttype/RigidTypes.h>
+#include <sofa/defaulttype/typeinfo/DataTypeInfo_RigidTypes.h>
 #include <sofa/defaulttype/typeinfo/DataTypeInfo_Vec.h>
 
 #include <vector>
