+++ modules/SofaBoundaryCondition/ProjectDirectionConstraint.inl
@@ -29,6 +29,7 @@
 #include <sofa/simulation/Simulation.h>
 #include <iostream>
 #include <SofaBaseTopology/TopologySubsetData.inl>
+#include <sofa/defaulttype/typeinfo/DataTypeInfo_Scalar.h>
 
 
 namespace sofa
