+++ modules/SofaConstraint/StickContactConstraint.cpp
@@ -19,6 +19,8 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
+#include <sofa/defaulttype/typeinfo/DataTypeInfo_Scalar.h>
+
 #include <SofaConstraint/StickContactConstraint.inl>
 #include <SofaMeshCollision/BarycentricContactMapper.h>
 #include <SofaMeshCollision/RigidContactMapper.h>
