+++ modules/SofaMiscMapping/BeamLinearMapping.inl
@@ -22,8 +22,6 @@
 #ifndef SOFA_COMPONENT_MAPPING_BEAMLINEARMAPPING_INL
 #define SOFA_COMPONENT_MAPPING_BEAMLINEARMAPPING_INL
 
-#include <sofa/defaulttype/typeinfo/DataTypeInfo_Vec.h>
-
 #include <SofaMiscMapping/BeamLinearMapping.h>
 #include <sofa/core/visual/VisualParams.h>
 #include <sofa/helper/io/SphereLoader.h>
