+++ modules/SofaMiscMapping/DistanceFromTargetMapping.cpp
@@ -21,6 +21,9 @@
 ******************************************************************************/
 #define SOFA_COMPONENT_MAPPING_DistanceFromTargetMapping_CPP
 
+
+#include <sofa/defaulttype/typeinfo/DataTypeInfo_RigidTypes.h>
+
 #include "DistanceFromTargetMapping.inl"
 #include <sofa/core/ObjectFactory.h>
