+++ modules/SofaMiscMapping/SofaMiscMapping_test/SubsetMultiMapping_test.cpp
@@ -24,6 +24,7 @@
 #include <sofa/defaulttype/VecTypes.h>
 //#include <sofa/defaulttype/RigidTypes.h>
 #include <SofaMiscMapping/SubsetMultiMapping.h>
+#include <sofa/defaulttype/typeinfo/DataTypeInfo_RigidTypes.h>
 
 
 namespace sofa {
