+++ modules/SofaMiscMapping/SquareMapping.cpp
@@ -23,6 +23,7 @@
 
 #include "SquareMapping.inl"
 #include <sofa/core/ObjectFactory.h>
+#include <sofa/defaulttype/typeinfo/DataTypeInfo_RigidTypes.h>
 
 namespace sofa
 {
