+++ modules/SofaMiscMapping/TubularMapping.h
@@ -27,6 +27,7 @@
 #include <sofa/core/behavior/MechanicalState.h>
 #include <sofa/defaulttype/RigidTypes.h>
 #include <sofa/defaulttype/typeinfo/DataTypeInfo_Vec.h>
+#include <sofa/defaulttype/typeinfo/DataTypeInfo_RigidTypes.h>
 #include <SofaGeneralSimpleFem/RadiusContainer.h>
 #include <vector>
